Trim whitespace from URI
If you form your Jekyll tag with whitespaces for formatting, those whitespaces are included in the `open` method call, which causes a `(URI::InvalidURIError)` error to be returned. Instead of removing the whitespace from the curly braces and produce something like: `{%github_repo_tag user/repo%}`, this commit removes the whitespace retroactively during the method call.